Add dynamic version information based on git tag
An issue existed where a default version existed for both the rpm
and library. This should not be the case, the versions of both the
rpm and the library should be ever increasing as the code is developed.
Tie the version of the code to the annotated git tags in the form of

Major.Minor.Hotfix.Build Number

The function to create an rpm was moved into the top layer makefile to
accommodate this change.

Signed-off-by: Nicholas Moulin <nicholas.w.moulin@intel.com>