Add ignore as option in create namespace flags
When creating a namespace, the required erase capabilities and
encryption capabilities of the underlying pool can be specified
on the command line. Add an option to use a pool regardless of
the erase capabilities or the encryption capabilities (as separate
options). This means that specifying 'No' for either option will
now require that the pool not support the corresponding feature.

Also make the new 'ignore' option the default value.

Signed-off-by: Kevin Degi <kevin.degi@intel.com>