Enforce interleave ways when specified
If the user specifies that they want a namespace created on
a non-interleaved App Direct pool, do not allow it to be created on
an interleaved pool. Likewise, if a user specifieds interleaved App
Direct, do not allow non-interleaved App Direct.

This includes the case where a user specifies the interleave sizes via
CIM. The specified interleave sizes are interpreted as a request for
an interleaved App Direct pool.

Signed-off-by: Kevin Degi <kevin.degi@intel.com>