Ignore bad DIMM serial number in goal interleave set
If a DIMM is removed from a valid configuration, its serial
number can no longer be correlated to a UID. This is not
corrupt data or a "bad config" - simply a sign that a DIMM
is missing. A missing DIMM's UID will be empty since we
can't look it up.

Signed-off-by: Kristin Jacque <kristin.jacque@intel.com>