Autotoolize.
Introduces pkg-config checks to test for invm-frameworks components,
and removes the manual makefiles since they would conflict entirely
with the generated ones made by automake.

This requires a version of invm-frameworks that has been autotooled
as well due to the reliance on pkgconfig data.

Passes `make distcheck`. The project does not yet create pkgconfig
files or install headers. It's not yet clear to me at this point
what headers constitute the API but this should be trivial to solve.

This has been tested only on Linux for now, but there are hooks
in place to make this easy to work with under windows with MINGW.