Introducing a simpler way to get properties from FIS commands
Eventually other places too. It relies on auto-generated code,
so changes are made in .py files and .template files and generated
into .c files.

Implement the first demo with getting properties from memory
information pages 0, 1, and 3 using the existing C++ api, but we're
shooting for using a new implementation of the CLI interface.

Signed-off-by: Nolan Hergert <nolan.hergert@intel.com>